requirements.txt
Django==1.11.6
PyYAML==3.12
gunicorn==19.7.1